requirements.txt
pymupdf
camelot-py
scikit-image
matplotlib
imagehash